Created Footer. Separated Header from body
